pyproject.toml
[project]
name = "yume"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel Sendzik"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "<4.0,>=3.13"
dependencies = [
    "fastapi>=0.116.1, <1.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
    "openai-agents (>=0.3.0,<0.4.0)",
    "matrix-nio>=0.25.2",
    "pytz (>=2025.2,<2026.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.1.3,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
